Replace 'by-id' drives in fstab
Syntax for '/dev/disk/by-id' has '-partX' for the partition behind the
drive paths. So we need to replace them here too, if found. Else the
script would try to format the underlying drive and not the raid array,
which would not work, because after joining the raid array the drive is
constantly busy.